ai_agent/pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.1"
description = ""
authors = ["Nir Bar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "^0.3.20"
langchain-openai = "^0.3.7"
langchain-community = "^0.3.19"
langgraph = "^0.3.2"
grandalf = "^0.8"
sendgrid = "^6.11.0"
streamlit = "^1.38.0"
serpapi = "^0.1.5"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
